moved logger and report modules to aop module
