+++ aop/build.gradle.kts
@@ -1,0 +1,5 @@
+allprojects {
+    dependencies {
+        implementation("org.aspectj:aspectjweaver:1.9.20.1")
+    }
+}
