+++ aop/logger/src/main/java/executor/service/aop/logger/aspect/LogMessage.java
@@ -1,0 +1,17 @@
+package executor.service.aop.logger.aspect;
+
+enum LogMessage {
+    EXECUTING_METHOD("Executing method {} in class {} with arguments {}"),
+    METHOD_EXECUTION_COMPLETED("Method {} execution completed in class {}"),
+    INVOCATION_TARGET_EXCEPTION("Method {} execution failed in class: {}");
+
+    private final String message;
+
+    LogMessage(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
