+++ aop/logger/src/main/java/executor/service/aop/logger/config/LoggerConfiguration.java
@@ -1,0 +1,14 @@
+package executor.service.aop.logger.config;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class LoggerConfiguration {
+    @Bean
+    public Logger logger() {
+        return LoggerFactory.getLogger("PROJECT_LOGGER");
+    }
+}
