+++ aop/report/build.gradle.kts
@@ -1,0 +1,6 @@
+dependencies {
+    implementation(project(":model"))
+    implementation(project(":execution"))
+    implementation(project(":redis:repository"))
+    implementation(project(":redis:queue"))
+}
