+++ execution/build.gradle.kts
@@ -1,6 +1,6 @@
 dependencies {
     implementation("org.seleniumhq.selenium:selenium-java:4.13.0")
-    implementation(project(":logger"))
+    implementation(project(":aop:logger"))
     implementation(project(":model"))
     implementation(project(":webdriver"))
     implementation(project(":logback"))
