+++ facade/build.gradle
@@ -2,7 +2,7 @@ dependencies {
     implementation(project(":execution"))
     implementation(project(":model"))
     implementation(project(":webdriver"))
-    implementation(project(":report"))
+    implementation(project(":aop:report"))
     implementation(project(":redis:queue"))
     implementation("org.seleniumhq.selenium:selenium-java:4.13.0")
