--- logger/build.gradle.kts
@@ -1,3 +1,0 @@
-dependencies {
-    implementation("org.aspectj:aspectjweaver:1.9.20.1")
-}
