--- logger/src/main/java/executor/service/logger/annotation/HandleException.java
@@ -1,11 +1,0 @@
-package executor.service.logger.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface HandleException {
-}
