+++ webdriver/build.gradle.kts
@@ -1,5 +1,5 @@
 dependencies {
     implementation("org.seleniumhq.selenium:selenium-java:4.13.0")
     implementation(project(":model"))
-    implementation(project(":logger"))
+    implementation(project(":aop:logger"))
 }
